FROMLIST: clk: rockchip: rk3288: prefer vdpu for vcodec clock source
Patch provided by Randy Li. The original commit message reads :

_______________
The RK3288 CRU system clock solution would suggest use
the vdpu clock source for the VPU(aclk_vpu and hclk_vpu).

Reading the registers of VPU(both VEPU and VDPU) would become all high
when the vepu is used as the clock source. It may be a bug in the SoC,
not sure whether it is fixed at RK3288W.

Signed-off-by: Randy Li <ayaka@soulik.info>
_______________

This also resolves a freeze when loading the OOT Video Codec driver

Change-Id: I6886e1b077a16c96997ad26f4da6e9a9cdf7313f
Signed-off-by: Myy <myy@miouyouyou.fr>