arm64: dts: rockchip: Add support for rk3399 sapphire SOM
Add support for the rk3399 sapphire SOM board.
This board works in a combination with the excavator main board.

Add support for the rk3399 sapphire SOM board.
This board works in a combination with the excavator main board.

This board have been sold as the rk3399 evaluation board for commercial customers.
You can get documents from below link:
http://opensource.rock-chips.com/wiki_Excavator_sapphire_board

Change-Id: Ifd536bc74c385877ba56490789617094f7b839b4
Signed-off-by: Jacob Chen <jacob-chen@iotwrt.com>